# api-gateway/docs/swagger.yaml
definitions:
...
  userproto.Courier:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
  userproto.LoginCourierRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  userproto.LoginCourierResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
  userproto.RegisterCourierRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
  userproto.RegisterCourierResponse:
    properties:
      message:
        type: string
    type: object
  userproto.RegisterReq:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  userproto.RegisterRes:
    properties:
      message:
        type: string
    type: object
  userproto.User:
    properties:
      email:
        type: string
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  userproto.UserReq:
    properties:
      code:
        type: integer
      email:
        type: string
    type: object
  userproto.UserRes:
    properties:
      id:
        type: integer
    type: object
info:
  contact: {}
  description: This is a sample server for a restaurant reservation system.
  title: Food Delivery
  version: "1.0"
paths:
  /couriers/:
    put:
      consumes:
      - application/json
      description: Update the details of an existing courier
      parameters:
      - description: Courier update request
        in: body
        name: courier
        required: true
        schema:
          $ref: '#/definitions/userproto.Courier'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userproto.Courier'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer Auth: []
      summary: Update courier details
      tags:
      - courier
  /couriers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing courier by courier ID
      parameters:
      - description: Courier ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userproto.RegisterCourierResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer Auth: []
      summary: Delete courier by ID
      tags:
      - courier
  /couriers/login:
    post:
      consumes:
      - application/json
      description: Authenticate a courier and return an access token
      parameters:
      - description: Courier login request
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/userproto.LoginCourierRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userproto.LoginCourierResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login courier
      tags:
      - courier
  /couriers/register:
    post:
      consumes:
      - application/json
      description: Create a new courier with the given details
      parameters:
      - description: Courier registration request
        in: body
        name: courier
        required: true
        schema:
          $ref: '#/definitions/userproto.RegisterCourierRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userproto.RegisterCourierResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register a new courier
      tags:
      - courier
  /couriers/verify-code:
    post:
      consumes:
      - application/json
      description: Verify the courier's code for registration or password reset
      parameters:
      - description: Verification request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/userproto.UserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userproto.UserRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Verify courier code
      tags:
      - courier
...
  /users/:
    put:
      consumes:
      - application/json
      description: Update the details of an existing user
      parameters:
      - description: User update request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/userproto.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userproto.RegisterRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer Auth: []
      summary: Update user details
      tags:
      - user
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing user by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userproto.RegisterRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer Auth: []
      summary: Delete user by ID
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Retrieve user details by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userproto.User'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer Auth: []
      summary: Get user by ID
      tags:
      - user
  /users/register:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User request body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/userproto.RegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userproto.RegisterRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new user
      tags:
      - user
  /users/verify-code:
    post:
      consumes:
      - application/json
      description: Verify the user's code for registration or password reset
      parameters:
      - description: Verification request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/userproto.UserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userproto.UserRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Verify user code
      tags:
      - user
securityDefinitions:
  Bearer:
    description: Enter the token in the format `Bearer {token}`
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"